Fix default values for component
The `escapeMarkup` opens jinn-toast up for a script injection attack. If escapeMarkup attribute is
not passed, it will default to the boolean true, which is _not_ equal to the string 'true'. Causing
it to be false and the text to eventually be set as `innerHtml`. So setting the text to anything
HTML (like a script tag) opens a vector for script injection.